+++ src/TitleDetails/ReceivedPiecesList/ReceivedPiecesList.js
@@ -2,20 +2,30 @@ import React from 'react';
 import PropTypes from 'prop-types';
 
 import PiecesList from '../PiecesList';
+import ReceivedPiecesActions from '../PiecesActions';
 
-const visibleColumns = ['title', 'caption', 'format', 'receivedDate'];
+const visibleColumns = ['title', 'caption', 'format', 'receivedDate', 'actions'];
 
-const ReceivedPiecesList = ({ pieces, title }) => {
+const ReceivedPiecesList = ({ pieces, title, onUnreceivePiece }) => {
+  const renderActions = (piece) => (
+    <ReceivedPiecesActions
+      onUnreceivePiece={onUnreceivePiece}
+      receivedPiece={piece}
+    />
+  );
+
   return (
     <PiecesList
       title={title}
       pieces={pieces}
       visibleColumns={visibleColumns}
+      renderActions={renderActions}
     />
   );
 };
 
 ReceivedPiecesList.propTypes = {
+  onUnreceivePiece: PropTypes.func.isRequired,
   pieces: PropTypes.arrayOf(PropTypes.object),
   title: PropTypes.string,
 };