+++ src/TitleDetails/TitleDetails.js
@@ -1,4 +1,4 @@
-import React, { useCallback } from 'react';
+import React, { useCallback, useState } from 'react';
 import PropTypes from 'prop-types';
 import {
   get,
@@ -44,6 +44,7 @@ const TitleDetails = ({
   onCheckIn,
   onClose,
   onEdit,
+  onUnreceivePiece,
   pieces,
   poLine,
   title,
@@ -51,6 +52,8 @@ const TitleDetails = ({
   const [expandAll, sections, toggleSection] = useAccordionToggle();
   const [isAcknowledgeNote, toggleAcknowledgeNote] = useModalToggle();
   const [isAddPieceModalOpened, toggleAddPieceModalOpened] = useModalToggle();
+  const [isUnreceiveConfirmation, toggleUnreceiveConfirmation] = useModalToggle();
+  const [pieceToUnreceive, setPieceToUnreceive] = useState();
   const receivingNote = get(poLine, 'details.receivingNote');
   const expectedPieces = pieces.filter(({ receivingStatus }) => receivingStatus === PIECE_STATUS.expected);
   const receivedPieces = sortBy(pieces.filter(
@@ -100,6 +103,22 @@ const TitleDetails = ({
     >
       <FormattedMessage id="ui-receiving.piece.button.addPiece" />
     </Button>
+  );
+
+  const confirmUnreceivePiece = useCallback(
+    () => {
+      onUnreceivePiece(pieceToUnreceive);
+      toggleUnreceiveConfirmation();
+    },
+    [pieceToUnreceive, onUnreceivePiece, toggleUnreceiveConfirmation],
+  );
+
+  const mountUnreceivePieceConfirmation = useCallback(
+    (piece) => {
+      setPieceToUnreceive(piece);
+      toggleUnreceiveConfirmation();
+    },
+    [toggleUnreceiveConfirmation, setPieceToUnreceive],
   );
 
   const lastMenu = (
@@ -177,6 +196,7 @@ const TitleDetails = ({
           label={TITLE_ACCORDION_LABELS.received}
         >
           <ReceivedPiecesList
+            onUnreceivePiece={mountUnreceivePieceConfirmation}
             pieces={receivedPieces}
             title={title.title}
           />
@@ -210,6 +230,18 @@ const TitleDetails = ({
           pieceFormatOptions={pieceFormatOptions}
         />
       )}
+
+      {isUnreceiveConfirmation && (
+        <ConfirmationModal
+          confirmLabel={<FormattedMessage id="ui-receiving.piece.actions.unreceive.confirm" />}
+          heading={<FormattedMessage id="ui-receiving.piece.actions.unreceive.heading" />}
+          id="unreceive-piece-confirmation"
+          message={<FormattedMessage id="ui-receiving.piece.actions.unreceive.message" />}
+          onCancel={toggleUnreceiveConfirmation}
+          onConfirm={confirmUnreceivePiece}
+          open
+        />
+      )}
     </Pane>
   );
 };
@@ -220,6 +252,7 @@ TitleDetails.propTypes = {
   onCheckIn: PropTypes.func.isRequired,
   onClose: PropTypes.func.isRequired,
   onEdit: PropTypes.func.isRequired,
+  onUnreceivePiece: PropTypes.func.isRequired,
   pieces: PropTypes.arrayOf(PropTypes.object),
   poLine: PropTypes.object.isRequired,
   title: PropTypes.object.isRequired,