+++ src/TitleDetails/utils.js
@@ -3,6 +3,8 @@ import {
   some,
 } from 'lodash';
 
+import { ITEM_STATUS } from './constants';
+
 export const checkInItems = (pieces, mutator) => {
   const linesMap = {};
 
@@ -59,3 +61,27 @@ export const getPOLLocationsForSelect = (locations, poLineLocations = []) => (
     );
   }).filter(Boolean)
 );
+
+export const unreceivePiece = (piece, mutator) => {
+  const { id, poLineId } = piece;
+  const item = {
+    itemStatus: ITEM_STATUS.onOrder,
+    pieceId: id,
+  };
+  const postData = {
+    toBeReceived: [{
+      poLineId,
+      received: 1,
+      receivedItems: [item],
+    }],
+    totalRecords: 1,
+  };
+
+  return mutator.POST(postData).then(({ receivingResults }) => {
+    if (some(receivingResults, ({ processedWithError }) => processedWithError > 0)) {
+      return Promise.reject(receivingResults);
+    }
+
+    return receivingResults;
+  });
+};